+++ gui/frontend/wailsjs/go/main/App.js
@@ -62,6 +62,10 @@ export function LoadFocusedModel(arg1, arg2) {
   return window['go']['main']['App']['LoadFocusedModel'](arg1, arg2);
 }
 
+export function LoadHuntLeads(arg1, arg2, arg3) {
+  return window['go']['main']['App']['LoadHuntLeads'](arg1, arg2, arg3);
+}
+
 export function LoadModel(arg1) {
   return window['go']['main']['App']['LoadModel'](arg1);
 }